requirements.txt
# Python >= 3.8
openai>=1.0.0
python-dotenv>=0.19.0
fastapi>=0.68.0
uvicorn>=0.15.0
websockets>=10.0 